+++ tests/test_signoff_flow.py
@@ -2,6 +2,8 @@ import unittest
 
 import mock
 
+from kinto.core.testing import FormattedErrorMixin
+from kinto.core.errors import ERRORS
 from .support import BaseWebTest, get_user_headers
 
 
@@ -75,13 +77,18 @@ class PostgresWebTest(BaseWebTest):
         return settings
 
 
-class CollectionStatusTest(PostgresWebTest, unittest.TestCase):
+class CollectionStatusTest(PostgresWebTest, FormattedErrorMixin, unittest.TestCase):
 
     def test_status_cannot_be_set_to_unknown_value(self):
-        self.app.patch_json(self.source_collection,
-                            {"data": {"status": "married"}},
-                            headers=self.headers,
-                            status=400)
+        resp = self.app.patch_json(self.source_collection,
+                                   {"data": {"status": "married"}},
+                                   headers=self.headers,
+                                   status=400)
+        self.assertFormattedError(response=resp,
+                                  code=400,
+                                  errno=ERRORS.INVALID_POSTED_DATA,
+                                  error="Bad Request",
+                                  message="Invalid status 'married'")
         resp = self.app.get(self.source_collection, headers=self.headers)
         assert resp.json["data"]["status"] == "work-in-progress"
 
@@ -282,7 +289,7 @@ class TrackingFieldsTest(PostgresWebTest, unittest.TestCase):
                               status=400)
 
 
-class UserGroupsTest(PostgresWebTest, unittest.TestCase):
+class UserGroupsTest(PostgresWebTest, FormattedErrorMixin, unittest.TestCase):
     def get_app_settings(self, extras=None):
         settings = super(UserGroupsTest, self).get_app_settings(extras)
         settings['signer.group_check_enabled'] = 'true'
@@ -311,10 +318,16 @@ class UserGroupsTest(PostgresWebTest, unittest.TestCase):
                           headers=self.headers)
 
     def test_only_editors_can_ask_to_review(self):
-        self.app.patch_json(self.source_collection,
-                            {"data": {"status": "to-review"}},
-                            headers=self.reviewer_headers,
-                            status=403)
+        resp = self.app.patch_json(self.source_collection,
+                                   {"data": {"status": "to-review"}},
+                                   headers=self.reviewer_headers,
+                                   status=403)
+        self.assertFormattedError(response=resp,
+                                  code=403,
+                                  errno=ERRORS.FORBIDDEN,
+                                  error="Forbidden",
+                                  message="Not in editors group")
+
         self.app.patch_json(self.source_collection,
                             {"data": {"status": "to-review"}},
                             headers=self.editor_headers)
@@ -324,10 +337,16 @@ class UserGroupsTest(PostgresWebTest, unittest.TestCase):
                             {"data": {"status": "to-review"}},
                             headers=self.editor_headers)
 
-        self.app.patch_json(self.source_collection,
-                            {"data": {"status": "to-sign"}},
-                            headers=self.editor_headers,
-                            status=403)
+        resp = self.app.patch_json(self.source_collection,
+                                   {"data": {"status": "to-sign"}},
+                                   headers=self.editor_headers,
+                                   status=403)
+        self.assertFormattedError(response=resp,
+                                  code=403,
+                                  errno=ERRORS.FORBIDDEN,
+                                  error="Forbidden",
+                                  message="Not in reviewers group")
+
         self.app.patch_json(self.source_collection,
                             {"data": {"status": "to-sign"}},
                             headers=self.reviewer_headers)